# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SpaceXStarlink API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PlanCategory = (typeof PlanCategory)[keyof typeof PlanCategory];

export const PlanCategory = {
  residential: "residential",
  roam: "roam",
  business: "business",
} as const;

export interface Plan {
  id: number;
  name: string;
  category: PlanCategory;
  speed: string;
  priceMonthly: number;
  features: string[];
  stripePriceId?: string;
  active: boolean;
  popular: boolean;
  description: string;
  hardwarePrice?: number;
  createdAt?: string;
}

export type CreatePlanBodyCategory =
  (typeof CreatePlanBodyCategory)[keyof typeof CreatePlanBodyCategory];

export const CreatePlanBodyCategory = {
  residential: "residential",
  roam: "roam",
  business: "business",
} as const;

export interface CreatePlanBody {
  name: string;
  category: CreatePlanBodyCategory;
  speed: string;
  priceMonthly: number;
  features: string[];
  stripePriceId?: string;
  popular?: boolean;
  description: string;
  hardwarePrice?: number;
}

export interface UpdatePlanBody {
  name?: string;
  speed?: string;
  priceMonthly?: number;
  features?: string[];
  stripePriceId?: string;
  active?: boolean;
  popular?: boolean;
  description?: string;
  hardwarePrice?: number;
}

export type SubscriptionStatus =
  (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus];

export const SubscriptionStatus = {
  active: "active",
  cancelled: "cancelled",
  past_due: "past_due",
  trialing: "trialing",
  incomplete: "incomplete",
} as const;

export interface Subscription {
  id: number;
  email: string;
  name?: string;
  planId: number;
  planName: string;
  planCategory: string;
  priceMonthly: number;
  stripeSubscriptionId?: string;
  stripeCustomerId?: string;
  status: SubscriptionStatus;
  address?: string;
  createdAt: string;
  cancelledAt?: string;
}

export interface SubscriptionList {
  subscriptions: Subscription[];
  total: number;
  page: number;
  limit: number;
}

export type UpdateSubscriptionBodyStatus =
  (typeof UpdateSubscriptionBodyStatus)[keyof typeof UpdateSubscriptionBodyStatus];

export const UpdateSubscriptionBodyStatus = {
  active: "active",
  cancelled: "cancelled",
  past_due: "past_due",
} as const;

export interface UpdateSubscriptionBody {
  status?: UpdateSubscriptionBodyStatus;
}

export interface CreateCheckoutSessionBody {
  planId: number;
  email: string;
  name: string;
  address?: string;
  successUrl?: string;
  cancelUrl?: string;
}

export interface CheckoutSession {
  sessionId: string;
  url: string;
}

export interface CheckoutSuccessResponse {
  success: boolean;
  subscription?: Subscription;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
}

export type AdminStatsPlanBreakdownItem = {
  planName: string;
  count: number;
  revenue: number;
};

export interface AdminStats {
  totalSubscriptions: number;
  activeSubscriptions: number;
  cancelledSubscriptions: number;
  monthlyRevenue: number;
  totalRevenue: number;
  newThisMonth: number;
  planBreakdown: AdminStatsPlanBreakdownItem[];
  recentSubscriptions: Subscription[];
}

export type RevenueStatsMonthlyItem = {
  month: string;
  revenue: number;
  subscriptions: number;
};

export interface RevenueStats {
  monthly: RevenueStatsMonthlyItem[];
}

export type CheckoutSuccessParams = {
  session_id?: string;
};

export type ListSubscriptionsParams = {
  status?: ListSubscriptionsStatus;
  page?: number;
  limit?: number;
};

export type ListSubscriptionsStatus =
  (typeof ListSubscriptionsStatus)[keyof typeof ListSubscriptionsStatus];

export const ListSubscriptionsStatus = {
  active: "active",
  cancelled: "cancelled",
  past_due: "past_due",
  all: "all",
} as const;