# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SpaceXStarlink API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials
...
  /admin/plans:
    get:
      operationId: adminListPlans
      tags: [admin]
      summary: Admin list all plans including inactive
      responses:
        "200":
          description: All plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: adminCreatePlan
      tags: [admin]
      summary: Create a new plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanBody"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /admin/plans/{id}:
    patch:
      operationId: adminUpdatePlan
      tags: [admin]
      summary: Update a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanBody"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: adminDeletePlan
      tags: [admin]
      summary: Delete (deactivate) a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
...
components:
  schemas:
...
    Plan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [residential, roam, business]
        speed:
          type: string
        priceMonthly:
          type: number
        features:
          type: array
          items:
            type: string
        stripePriceId:
          type: string
        active:
          type: boolean
        popular:
          type: boolean
        description:
          type: string
        hardwarePrice:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - speed
        - priceMonthly
        - features
        - active
        - popular
        - description
    CreatePlanBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [residential, roam, business]
        speed:
          type: string
        priceMonthly:
          type: number
        features:
          type: array
          items:
            type: string
        stripePriceId:
          type: string
        popular:
          type: boolean
        description:
          type: string
        hardwarePrice:
          type: number
      required:
        - name
        - category
        - speed
        - priceMonthly
        - features
        - description
    UpdatePlanBody:
      type: object
      properties:
        name:
          type: string
        speed:
          type: string
        priceMonthly:
          type: number
        features:
          type: array
          items:
            type: string
        stripePriceId:
          type: string
        active:
          type: boolean
        popular:
          type: boolean
        description:
          type: string
        hardwarePrice:
          type: number
...
    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
      required:
        - success
        - token
...